// frontend/src/client/models/StudentCreate.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type StudentCreate = {
    full_name: string;
    city?: (string | null);
    responsible_adult_full_name?: (string | null);
    responsible_adult_phone_number?: (string | null);
    notes?: (string | null);
    groups: Array<number>;
};