// frontend/src/client/models/GroupOut.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { GroupStudentLink } from './GroupStudentLink';

export type GroupOut = {
    name: string;
    description: string;
    id: number;
    student_links: Array<GroupStudentLink>;
};